refactor: simplify classes
fix: handle weeks


lint: cleanup and no more pylint

fix: use more robust pandas datetime64 check


test: simplify cast endpoints
